requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
ultralytics>=8.0.0
gTTS>=2.3.2
pygame>=2.5.0
Pillow>=10.0.0
easyocr>=1.7.0
face-recognition>=1.3.0
customtkinter>=5.2.0